27/11 Mail & merge code

+++ src/main/java/sd_009/bookstore/entity/receipt/ReceiptDetail.java
@@ -25,7 +25,7 @@ public class ReceiptDetail extends AuditableEntity {
     @JoinColumn
     private BookDetail bookCopy;
 
-    private Double pricePerUnit;
+    private Long pricePerUnit;
 
     private Long quantity;
